requirement.txt
allure-behave
allure-pytest
allure-python-commons
Appium-Python-Client
attrs
beautifulsoup4
behave-html-formatter
boto3
botocore
certifi
cffi
charset-normalizer
colorama
cryptography
execnet
flake8
h11
idna
iniconfig
Jinja2
jmespath
Mako
MarkupSafe
mccabe
outcome
packaging
parse
parse-type
pluggy
psutil
pybase64
pycodestyle
pycparser
pyflakes
PyHamcrest
pyOpenSSL
PySocks
pytest
pytest-bdd
pytest-html
pytest-metadata
pytest-xdist
python-dateutil
python-dotenv
pytz
requests
roku
s3transfer
selenium
six
sniffio
sortedcontainers
soupsieve
trio
trio-websocket
typing_extensions
tzdata
tzlocal
urllib3
urllib3-secure-extra
webdriver-manager
websocket-client
wget
wsproto
zipp
pyautogui
pyperclip